APIs/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
